refactor: create `Auth` folder for auth-related pages
