+++ src/pages/Auth/index.js
@@ -1,0 +1,3 @@
+export { default as Login } from './login';
+export { default as Register } from './register';
+export { default as ResetPassword } from './resetPassword';
